Feat: 맛집 API 수정으로 인해 타입 수정 및 Review 관련 타입 분리
